Projeto Finalizado
Funcionalidade 2 e 3 implementada, juntamente com refatoramento de código.
