+++ CasaDoCodigo/Models/Categoria.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace CasaDoCodigo.Models
+{
+    public class Categoria : BaseModel
+    {
+    
+        public Categoria(string nome)
+        {
+            this.Nome = nome;
+        }
+        public Categoria()
+        {
+
+        }
+        [Required]
+        public String Nome { get; set; }
+
+
+    }
+
+}
