+++ CasaDoCodigo/Models/modelo.cs
@@ -1,162 +1,12 @@
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Runtime.Serialization;
 
 namespace CasaDoCodigo.Models
 {
-   
+
     public abstract class BaseModel
     {
         [DataMember]
         public int Id { get; protected set; }
     }
-    public class Categoria : BaseModel
-    {
-    
-        public Categoria(string nome)
-        {
-            this.Nome = nome;
-        }
-        public Categoria()
-        {
 
-        }
-        [Required]
-        public String Nome { get; set; }
-
-
-    }
-
-    public class Produto : BaseModel
-    {
-        public Produto()
-        {
-
-        }
-
-        [Required]
-        [DataMember]
-        public Categoria Categoria { get;  private set; }
-        [Required]
-        public string Codigo { get; private set; }
-        [Required]
-        public string Nome { get; private set; }
-        [Required]
-        public decimal Preco { get; private set; }
-        
-        public Produto( string codigo, string nome, decimal preco)
-        {
-         
-            this.Codigo = codigo;
-            this.Nome = nome;
-            this.Preco = preco;   
-        }
-        public Produto(string codigo, string nome, decimal preco, Categoria categoria)
-        {
-
-            this.Codigo = codigo;
-            this.Nome = nome;
-            this.Preco = preco;
-            this.Categoria = categoria;
-        }
-
-
-    }
-
-    public class Cadastro : BaseModel
-    {
-        public Cadastro()
-        {
-        }
-
-        public virtual Pedido Pedido { get; set; }
-        [MinLength(5, ErrorMessage = "Nome deve ter no mínimo 5 caracteres")]
-        [MaxLength(50, ErrorMessage = "Nome deve ter no máximo 50 caracteres")]
-        [Required(ErrorMessage = "Nome é obrigatório")]
-        public string Nome { get; set; } = "";
-        [Required(ErrorMessage = "Email é obrigatório")]
-        public string Email { get; set; } = "";
-        [Required(ErrorMessage = "Telefone é obrigatório")]
-        public string Telefone { get; set; } = "";
-        [Required(ErrorMessage = "Endereco é obrigatório")]
-        public string Endereco { get; set; } = "";
-        [Required(ErrorMessage = "Complemento é obrigatório")]
-        public string Complemento { get; set; } = "";
-        [Required(ErrorMessage = "Bairro é obrigatório")]
-        public string Bairro { get; set; } = "";
-        [Required(ErrorMessage = "Municipio é obrigatório")]
-        public string Municipio { get; set; } = "";
-        [Required(ErrorMessage = "UF é obrigatório")]
-        public string UF { get; set; } = "";
-        [Required(ErrorMessage = "CEP é obrigatório")]
-        public string CEP { get; set; } = "";
-
-        internal void Update(Cadastro novoCadastro)
-        {
-            this.Bairro = novoCadastro.Bairro;
-            this.CEP = novoCadastro.CEP;
-            this.Complemento = novoCadastro.Complemento;
-            this.Email = novoCadastro.Email;
-            this.Endereco = novoCadastro.Endereco;
-            this.Municipio = novoCadastro.Municipio;
-            this.Nome = novoCadastro.Nome;
-            this.Telefone = novoCadastro.Telefone;
-            this.UF = novoCadastro.UF;
-        }
-    }
-
-    [DataContract]
-    public class ItemPedido : BaseModel
-    {
-        [Required]
-        [DataMember]
-        public Pedido Pedido { get; private set; }
-        [Required]
-        [DataMember]
-        public Produto Produto { get; private set; }
-        [Required]
-        [DataMember]
-        public int Quantidade { get; private set; }
-        [Required]
-        [DataMember]
-        public decimal PrecoUnitario { get; private set; }
-        [DataMember]
-        public decimal Subtotal => Quantidade * PrecoUnitario;
-
-        public ItemPedido()
-        {
-
-        }
-
-        public ItemPedido(Pedido pedido, Produto produto, int quantidade, decimal precoUnitario)
-        {
-            Pedido = pedido;
-            Produto = produto;
-            Quantidade = quantidade;
-            PrecoUnitario = precoUnitario;
-        }
-
-        internal void AtualizaQuantidade(int quantidade)
-        {
-            Quantidade = quantidade;
-        }
-    }
-
-    public class Pedido : BaseModel
-    {
-        public Pedido()
-        {
-            Cadastro = new Cadastro();
-        }
-
-        public Pedido(Cadastro cadastro)
-        {
-            Cadastro = cadastro;
-        }
-
-        public List<ItemPedido> Itens { get; private set; } = new List<ItemPedido>();
-        [Required]
-        public virtual Cadastro Cadastro { get; private set; }
-    }
 }
