+++ CasaDoCodigo/Repositories/IProdutoRepository.cs
@@ -1,4 +1,5 @@
 ﻿using CasaDoCodigo.Models;
+using CasaDoCodigo.Models.ViewModels;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -8,5 +9,6 @@ namespace CasaDoCodigo.Repositories
     {
         Task SaveProdutos(List<Livro> livros);
         IList<Produto> GetProdutos();
+        Task<BuscaProdutosViewModel> GetProdutosAsync(string pesquisa);
     }
 }
